// utils/payhere/types/payhere.types.ts
export interface PayHerePaymentRequest {
    [key: string]: any;
    merchant_id: string;
    return_url: string;
    cancel_url: string;
    notify_url: string;
    first_name: string;
    last_name: string;
    email: string;
    phone: string;
    address: string;
    city: string;
    country: string;
    order_id: string;
    items: string;
    currency: string;
    recurrence: string;
    duration: string;
    amount: number | string;
    hash: string;
    delivery_address?: string;
    delivery_city?: string;
    delivery_country?: string;
    item_name_1?: string;
    item_number_1?: string;
    amount_1?: number;
    quantity_1?: number;
    item_name_2?: string;
    item_number_2?: string;
    amount_2?: number;
    quantity_2?: number;
    platform?: string;
    custom_1?: string;
    custom_2?: string;
    startup_fee?: number;

}